extracted appender and dao to logback module
