+++ executor-service/build.gradle.kts
@@ -1,8 +1,7 @@
 dependencies {
-    implementation("com.h2database:h2")
     implementation("com.squareup.okhttp3:okhttp:4.11.0")
     implementation("org.seleniumhq.selenium:selenium-java:4.13.0")
     implementation("org.aspectj:aspectjweaver:1.9.20.1")
-    implementation("org.postgresql:postgresql:42.6.0")
     implementation("com.fasterxml.jackson.core:jackson-databind:2.15.2")
+    implementation(project(":logback"))
 }
