+++ logback/src/main/java/executor/service/logback/dao/manager/db/LogDatabaseManager.java
@@ -1,13 +1,13 @@
-package executor.service.appender.manager.db;
+package executor.service.logback.dao.manager.db;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.classic.spi.IThrowableProxy;
 import ch.qos.logback.classic.spi.StackTraceElementProxy;
-import executor.service.appender.manager.LogStorageManager;
-import executor.service.exception.logstorage.ConnectionFailedException;
-import executor.service.exception.logstorage.DisconnectionFailedException;
-import executor.service.exception.logstorage.LogEventSaveException;
-import executor.service.exception.logstorage.StackTraceSaveException;
+import executor.service.logback.dao.manager.LogStorageManager;
+import executor.service.logback.exception.logstorage.ConnectionFailedException;
+import executor.service.logback.exception.logstorage.DisconnectionFailedException;
+import executor.service.logback.exception.logstorage.LogEventSaveException;
+import executor.service.logback.exception.logstorage.StackTraceSaveException;
 
 import java.sql.*;
 import java.util.ArrayList;
