+++ logback/src/main/java/executor/service/logback/exception/logstorage/ConnectionFailedException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.logstorage;
+package executor.service.logback.exception.logstorage;
 
 public class ConnectionFailedException extends RuntimeException {
     public ConnectionFailedException(Throwable cause) {
