+++ logback/src/main/java/executor/service/logback/exception/logstorage/DisconnectionFailedException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.logstorage;
+package executor.service.logback.exception.logstorage;
 
 public class DisconnectionFailedException extends RuntimeException {
     public DisconnectionFailedException(Throwable cause) {
