+++ logback/src/main/java/executor/service/logback/exception/logstorage/StackTraceSaveException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.logstorage;
+package executor.service.logback.exception.logstorage;
 
 public class StackTraceSaveException extends RuntimeException {
     public StackTraceSaveException(Throwable cause) {
