+++ logback/src/main/resources/logback.xml
@@ -5,10 +5,10 @@
             <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} %-5level [%thread] %msg%n</pattern>
         </encoder>
     </appender>
-    <appender name="db" class="executor.service.appender.LogStorageAppender">
-        <logStorageManager class="executor.service.appender.manager.db.LogDatabaseManager">
-            <connectionProvider class="executor.service.appender.manager.db.ConnectionProviderImpl">
-                <url>jdbc:postgresql://postgres:5432/worker-logs</url>
+    <appender name="db" class="executor.service.logback.appender.LogStorageAppender">
+        <logStorageManager class="executor.service.logback.dao.manager.db.LogDatabaseManager">
+            <connectionProvider class="executor.service.logback.dao.manager.db.ConnectionProviderImpl">
+                <url>jdbc:postgresql://localhost:5432/worker-logs</url>
                 <username>root</username>
                 <password>root</password>
             </connectionProvider>
