+++ settings.gradle.kts
@@ -1,2 +1,3 @@
 rootProject.name = "worker_microservice"
 include("executor-service")
+include("logback")
